Redimension de ventana a pantalla

+++ ProyectoFiltros/B&W_Filter.h
@@ -12,7 +12,7 @@ void BnW_Filter(Mat img) {
     for (int y = 0; y < img.rows; y++) {
         for (int x = 0; x < img.cols; x++) {
          Vec3b color = img.at<Vec3b>(Point(x, y));
-         int graypixel = color[0] * 0.299 + color[1] * 0.587 + color[2] * 0.114;
+         double graypixel = color[0] * 0.299 + color[1] * 0.587 + color[2] * 0.114;
          color[0] = graypixel;
          color[1] = graypixel;
          color[2] = graypixel;
